[Android] Use foreground activity detection for intent delivery decision
Previously, we relied on client ping to decide whether to send an intent to an activity or show a notification if no activity was in the foreground. This worked because activities would unbind when backgrounded. With components like the 'Conduit state service' staying bound, pings can succeed even if all activities are backgrounded. To address this, we now use a method relying on ActivityManager.getRunningAppProcesses to reliably check if the app is in the foreground.